restructure - restructure the ui folders and files - group common components together and reuse
